reformat cli to use argparse for parsing args. restructure functions into submodule and files update pyproject.toml to reflect changes
