+++ src/swingmusic/args/__init__.py
@@ -1,0 +1,2 @@
+"""functions for handling arguments"""
+
